+++ sdk/batch/azure-mgmt-batch/azure/mgmt/batch/aio/operations/_certificate_operations.py
@@ -77,7 +77,7 @@ class CertificateOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-09-01"
+        api_version = "2021-01-01"
         accept = "application/json"
 
         def prepare_request(next_link=None):
@@ -135,7 +135,7 @@ class CertificateOperations:
         )
     list_by_batch_account.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Batch/batchAccounts/{accountName}/certificates'}  # type: ignore
 
-    async def _create_initial(
+    async def create(
         self,
         resource_group_name: str,
         account_name: str,
@@ -145,17 +145,41 @@ class CertificateOperations:
         if_none_match: Optional[str] = None,
         **kwargs
     ) -> "_models.Certificate":
+        """Creates a new certificate inside the specified account.
+
+        :param resource_group_name: The name of the resource group that contains the Batch account.
+        :type resource_group_name: str
+        :param account_name: The name of the Batch account.
+        :type account_name: str
+        :param certificate_name: The identifier for the certificate. This must be made up of algorithm
+         and thumbprint separated by a dash, and must match the certificate data in the request. For
+         example SHA1-a3d1c5.
+        :type certificate_name: str
+        :param parameters: Additional parameters for certificate creation.
+        :type parameters: ~azure.mgmt.batch.models.CertificateCreateOrUpdateParameters
+        :param if_match: The entity state (ETag) version of the certificate to update. A value of "*"
+         can be used to apply the operation only if the certificate already exists. If omitted, this
+         operation will always be applied.
+        :type if_match: str
+        :param if_none_match: Set to '*' to allow a new certificate to be created, but to prevent
+         updating an existing certificate. Other values will be ignored.
+        :type if_none_match: str
+        :keyword callable cls: A custom type or function that will be passed the direct response
+        :return: Certificate, or the result of cls(response)
+        :rtype: ~azure.mgmt.batch.models.Certificate
+        :raises: ~azure.core.exceptions.HttpResponseError
+        """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.Certificate"]
         error_map = {
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-09-01"
+        api_version = "2021-01-01"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
         # Construct URL
-        url = self._create_initial.metadata['url']  # type: ignore
+        url = self.create.metadata['url']  # type: ignore
         path_format_arguments = {
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
             'accountName': self._serialize.url("account_name", account_name, 'str', max_length=24, min_length=3, pattern=r'^[a-zA-Z0-9]+$'),
@@ -196,99 +220,7 @@ class CertificateOperations:
             return cls(pipeline_response, deserialized, response_headers)
 
         return deserialized
-    _create_initial.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Batch/batchAccounts/{accountName}/certificates/{certificateName}'}  # type: ignore
-
-    async def begin_create(
-        self,
-        resource_group_name: str,
-        account_name: str,
-        certificate_name: str,
-        parameters: "_models.CertificateCreateOrUpdateParameters",
-        if_match: Optional[str] = None,
-        if_none_match: Optional[str] = None,
-        **kwargs
-    ) -> AsyncLROPoller["_models.Certificate"]:
-        """Creates a new certificate inside the specified account.
-
-        :param resource_group_name: The name of the resource group that contains the Batch account.
-        :type resource_group_name: str
-        :param account_name: The name of the Batch account.
-        :type account_name: str
-        :param certificate_name: The identifier for the certificate. This must be made up of algorithm
-         and thumbprint separated by a dash, and must match the certificate data in the request. For
-         example SHA1-a3d1c5.
-        :type certificate_name: str
-        :param parameters: Additional parameters for certificate creation.
-        :type parameters: ~azure.mgmt.batch.models.CertificateCreateOrUpdateParameters
-        :param if_match: The entity state (ETag) version of the certificate to update. A value of "*"
-         can be used to apply the operation only if the certificate already exists. If omitted, this
-         operation will always be applied.
-        :type if_match: str
-        :param if_none_match: Set to '*' to allow a new certificate to be created, but to prevent
-         updating an existing certificate. Other values will be ignored.
-        :type if_none_match: str
-        :keyword callable cls: A custom type or function that will be passed the direct response
-        :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :keyword polling: True for ARMPolling, False for no polling, or a
-         polling object for personal polling strategy
-        :paramtype polling: bool or ~azure.core.polling.AsyncPollingMethod
-        :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
-        :return: An instance of AsyncLROPoller that returns either Certificate or the result of cls(response)
-        :rtype: ~azure.core.polling.AsyncLROPoller[~azure.mgmt.batch.models.Certificate]
-        :raises ~azure.core.exceptions.HttpResponseError:
-        """
-        polling = kwargs.pop('polling', True)  # type: Union[bool, AsyncPollingMethod]
-        cls = kwargs.pop('cls', None)  # type: ClsType["_models.Certificate"]
-        lro_delay = kwargs.pop(
-            'polling_interval',
-            self._config.polling_interval
-        )
-        cont_token = kwargs.pop('continuation_token', None)  # type: Optional[str]
-        if cont_token is None:
-            raw_result = await self._create_initial(
-                resource_group_name=resource_group_name,
-                account_name=account_name,
-                certificate_name=certificate_name,
-                parameters=parameters,
-                if_match=if_match,
-                if_none_match=if_none_match,
-                cls=lambda x,y,z: x,
-                **kwargs
-            )
-
-        kwargs.pop('error_map', None)
-        kwargs.pop('content_type', None)
-
-        def get_long_running_output(pipeline_response):
-            response_headers = {}
-            response = pipeline_response.http_response
-            response_headers['ETag']=self._deserialize('str', response.headers.get('ETag'))
-            deserialized = self._deserialize('Certificate', pipeline_response)
-
-            if cls:
-                return cls(pipeline_response, deserialized, response_headers)
-            return deserialized
-
-        path_format_arguments = {
-            'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'accountName': self._serialize.url("account_name", account_name, 'str', max_length=24, min_length=3, pattern=r'^[a-zA-Z0-9]+$'),
-            'certificateName': self._serialize.url("certificate_name", certificate_name, 'str', max_length=45, min_length=5, pattern=r'^[\w]+-[\w]+$'),
-            'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
-        }
-
-        if polling is True: polling_method = AsyncARMPolling(lro_delay, path_format_arguments=path_format_arguments,  **kwargs)
-        elif polling is False: polling_method = AsyncNoPolling()
-        else: polling_method = polling
-        if cont_token:
-            return AsyncLROPoller.from_continuation_token(
-                polling_method=polling_method,
-                continuation_token=cont_token,
-                client=self._client,
-                deserialization_callback=get_long_running_output
-            )
-        else:
-            return AsyncLROPoller(self._client, raw_result, get_long_running_output, polling_method)
-    begin_create.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Batch/batchAccounts/{accountName}/certificates/{certificateName}'}  # type: ignore
+    create.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Batch/batchAccounts/{accountName}/certificates/{certificateName}'}  # type: ignore
 
     async def update(
         self,
@@ -324,7 +256,7 @@ class CertificateOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-09-01"
+        api_version = "2021-01-01"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
@@ -382,7 +314,7 @@ class CertificateOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-09-01"
+        api_version = "2021-01-01"
         accept = "application/json"
 
         # Construct URL
@@ -478,7 +410,7 @@ class CertificateOperations:
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
         }
 
-        if polling is True: polling_method = AsyncARMPolling(lro_delay, path_format_arguments=path_format_arguments,  **kwargs)
+        if polling is True: polling_method = AsyncARMPolling(lro_delay, lro_options={'final-state-via': 'location'}, path_format_arguments=path_format_arguments,  **kwargs)
         elif polling is False: polling_method = AsyncNoPolling()
         else: polling_method = polling
         if cont_token:
@@ -519,7 +451,7 @@ class CertificateOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-09-01"
+        api_version = "2021-01-01"
         accept = "application/json"
 
         # Construct URL
@@ -592,7 +524,7 @@ class CertificateOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2020-09-01"
+        api_version = "2021-01-01"
         accept = "application/json"
 
         # Construct URL