+++ sdk/core/azure-core/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Release History
 
-## 1.10.1 (Unreleased)
+## 1.11.0 (Unreleased)
+
+### Features
+
+- Add `raise_for_status` method onto `HttpResponse`. Calling `response.raise_for_status()` on a response with an error code
+will raise an `HttpResponseError`. Calling it on a good response will do nothing  #16399
 
 
 ## 1.10.0 (2021-01-11)